Implementa loops e validações nas operações de conta
- Adiciona submenus em loop para repetir operações antes de voltar ao menu principal
- Troca valores monetários de `float` para `decimal`
- Valida titular obrigatório, limite não negativo e aniversário da poupança entre 1 e 31
- Simplifica a leitura de texto e valores decimais no console
- Documenta no README as decisões sobre `decimal` e aniversário da poupança

+++ Models/ContaCorrente.cs
@@ -6,11 +6,11 @@ namespace conta_bancaria_csharp.Models;
 /// </summary>
 public class ContaCorrente : Conta
 {
-    private float limite;
+    private decimal limite;
 
-    public float getLimite() { return limite; }
+    public decimal getLimite() { return limite; }
 
-    public void setLimite(float limite)
+    public void setLimite(decimal limite)
     {
         this.limite = limite;
     }
@@ -23,7 +23,7 @@ public class ContaCorrente : Conta
     /// <param name="tipo">Tipo da conta.</param>
     /// <param name="titular">Nome do titular da conta.</param>
     /// <param name="limite">Limite disponível para a conta corrente.</param>
-    public ContaCorrente(int tipo, string titular, float limite)
+    public ContaCorrente(int tipo, string titular, decimal limite)
         : base(tipo, titular)
     {
         setLimite(limite);
